Move PoolStatusTable to components folder and remove duplicate file
